add dashboard wrapper to fix suspense error
